+++ libs/container/src/ghostos_container.py
@@ -3,6 +3,7 @@ import inspect
 from abc import ABCMeta, abstractmethod
 from typing import Type, Dict, TypeVar, Callable, Set, Optional, List, Generic, Any, Union, Iterable
 from typing import get_args, get_origin, ClassVar
+import warnings
 
 __all__ = [
     "Container", "IoCContainer",
@@ -181,21 +182,16 @@ class Container(IoCContainer):
     bloodline: List[str]
 
     def __init__(self, parent: Optional[Container] = None, *, name: str = "", inherit: bool = True):
+        """
+        :param parent: parent container
+        :param name: name of the container
+        :param inherit: inherit the registrar from the parent container if given.
+        """
         self.bloodline = []
         # container extended by children container
-        if parent is not None:
-            if not isinstance(parent, Container):
-                raise AttributeError("container can only initialized with parent Container")
-            if parent is self:
-                raise AttributeError("container's parent must not be itself")
-        self.parent: Optional[Container] = parent
-        if isinstance(self.parent, Container):
-            bloodline = self.parent.bloodline.copy()
-            bloodline.append(name)
-        else:
-            bloodline = [name]
-        self.bloodline: List[str] = bloodline
-
+        self.parent: Optional[Container] = None
+        self.name = name
+        self.bloodline = [name]
         # global singletons.
         self._instances: Dict[Any, Any] = {}
         self._factory: Dict[Any, Factory] = {}
@@ -208,10 +204,27 @@ class Container(IoCContainer):
         self._is_shutdown: bool = False
         self._shutdown: List[Callable[[], None]] = []
         self._making_count: int = 0
-        if inherit and parent is not None:
-            self._inherit(parent)
+        # set parent now.
+        if parent is not None:
+            self.set_parent(parent, inherit)
 
         Container.instance_count += 1
+
+    def set_parent(self, parent: Container, shutdown: bool = True, inherit: bool = True) -> None:
+        if not isinstance(parent, Container):
+            raise AttributeError("container can only initialized with parent Container")
+        if parent is self:
+            raise AttributeError("container's parent must not be itself")
+        self.parent = parent
+        bloodline = self.parent.bloodline.copy()
+        bloodline.append(self.name)
+        self.bloodline = bloodline
+
+        if shutdown:
+            # when parent shutdown, shutdown self
+            parent.add_shutdown(self.shutdown)
+        if inherit and self.parent is not None:
+            self._inherit(self.parent)
 
     def _inherit(self, parent: Container):
         """
@@ -233,10 +246,6 @@ class Container(IoCContainer):
         if self._bootstrapper:
             for b in self._bootstrapper:
                 b.bootstrap(self)
-        for provider in self._providers.values():
-            # some bootstrapper provider may be override
-            if isinstance(provider, Bootstrapper):
-                provider.bootstrap(self)
 
     def add_shutdown(self, shutdown: Callable):
         self._shutdown.append(shutdown)
@@ -273,10 +282,16 @@ class Container(IoCContainer):
         """
         self._check_destroyed()
         # 进行初始化.
-        if not self._bootstrapped:
+        has_bootstrapper = len(self._bootstrapper)
+        if not self._bootstrapped and has_bootstrapper > 0:
             caller_info = get_caller_info(4)
-            # warnings.warn("container is not bootstrapped before using: %s" % (caller_info,))
-            self.bootstrap()
+            warnings.warn(
+                "container with %d bootstrapper have not bootstrapped before using: %s" % (
+                    has_bootstrapper, caller_info
+                ),
+                UserWarning,
+            )
+            # self.bootstrap()
 
         # get bound instance
         got = self._instances.get(abstract, None)
@@ -359,9 +374,9 @@ class Container(IoCContainer):
         for alias in provider.aliases():
             if alias not in self._bound:
                 self._bind_alias(alias, contract)
-        if isinstance(provider, Bootstrapper) and self._bootstrapped:
+        if isinstance(provider, Bootstrapper):
             # 添加 bootstrapper.
-            provider.bootstrap(self)
+            self.add_bootstrapper(provider)
 
     def _bind_alias(self, alias: Any, contract: Any) -> None:
         self._aliases[alias] = contract
@@ -381,8 +396,10 @@ class Container(IoCContainer):
         :return:
         """
         self._check_destroyed()
-        if not self._bootstrapped:
-            self._bootstrapper.append(bootstrapper)
+        self._bootstrapper.append(bootstrapper)
+        if self._bootstrapped:
+            # add bootstrapper and run it immediately
+            bootstrapper.bootstrap(self)
 
     def fetch(self, abstract: Type[INSTANCE], strict: bool = False) -> Optional[INSTANCE]:
         """
@@ -538,7 +555,6 @@ class Container(IoCContainer):
         del self._providers
         del self._bound
         del self._bootstrapper
-        del self._bootstrapped
         del self._aliases
         Container.instance_count -= 1
 
@@ -640,9 +656,6 @@ class ProviderAdapter(Generic[INSTANCE], Provider[INSTANCE]):
 
     def contract(self) -> Type[INSTANCE]:
         return self._contract_type
-
-    def inheritable(self) -> bool:
-        return False
 
     def factory(self, con: Container) -> Optional[INSTANCE]:
         return self._factory(con)