+++ libs/ghostos/ghostos/framework/messengers/defaults.py
@@ -33,6 +33,7 @@ class DefaultMessenger(Messenger):
         self._stage = stage
         self._destroyed = False
         self._output_pipes = output_pipes
+        self._buffering: Optional[Message] = None
         self.finish_reason = None
 
     def flush(self) -> Tuple[List[Message], List[FunctionCaller]]:
@@ -56,6 +57,12 @@ class DefaultMessenger(Messenger):
             # 非 function call 类型但也可以有 caller.
             elif message.callers:
                 callers.extend(message.callers)
+        # if buffering is not None, means interrupted.
+        if self._buffering is not None:
+            self.finish_reason = "interrupt"
+            messages.append(self._buffering.as_tail(copy=False))
+            callers = []
+
         self.destroy()
         return messages, callers
 
@@ -75,35 +82,29 @@ class DefaultMessenger(Messenger):
     def send(self, messages: Iterable[Message]) -> bool:
         messages = self.buffer(messages)
         if self._upstream is not None:
-            if self._output_pipes:
-                # set output pipes.
-                messages = run_pipeline(self._output_pipes, messages)
-
             return self._upstream.send(messages)
         list(messages)
         return True
 
     def buffer(self, messages: Iterable[Message]) -> Iterable[Message]:
         messages = SequencePipe().across(messages)
-        for item in messages:
-            # add message info
-            if item.is_complete() or item.is_head():
-                if not item.name and MessageType.is_text(item):
-                    item.name = self._assistant_name
-                if not item.stage:
-                    item.stage = self._stage
-                if not item.role:
-                    item.role = self._role
-            # create buffer in case upstream is cancel
-            if item.is_complete():
-                # add payload to complete one
-                if self._payloads:
-                    for payload in self._payloads:
-                        payload.set_payload_if_none(item)
+        if self._output_pipes:
+            # set output pipes.
+            messages = run_pipeline(self._output_pipes, messages)
 
+        for item in messages:
+            # update buffering.
+
+            item = self.wrap(item)
+            if item.is_complete():
                 # buffer outputs
                 self._sent_message_ids.append(item.msg_id)
                 self._sent_messages[item.msg_id] = item
+                self._buffering = None
+            elif self._buffering is None:
+                self._buffering = item.as_head(copy=True)
+            elif patched := self._buffering.patch(item):
+                self._buffering = patched
 
             # skip chunk
             if self._upstream and self._upstream.completes_only() and not item.is_complete():
@@ -111,6 +112,25 @@ class DefaultMessenger(Messenger):
             if item.is_complete() and item.finish_reason:
                 self.finish_reason = item.finish_reason
             yield item
+        # if sent all, the buffering is None.
+        self._buffering = None
+
+    def wrap(self, item: Message) -> Message:
+        # add message info
+        if item.is_complete() or item.is_head():
+            if not item.name and MessageType.is_text(item):
+                item.name = self._assistant_name
+            if not item.stage:
+                item.stage = self._stage
+            if not item.role:
+                item.role = self._role
+        # create buffer in case upstream is cancel
+        if item.is_complete():
+            # add payload to complete one
+            if self._payloads:
+                for payload in self._payloads:
+                    payload.set_payload_if_none(item)
+        return item
 
     def completes_only(self) -> bool:
         return self._upstream is not None and self._upstream.completes_only()