+++ libs/ghostos/ghostos/scripts/cli/run_console.py
@@ -1,5 +1,5 @@
 from ghostos.abcd import Ghost
-from ghostos.abcd.utils import get_module_magic_shell_providers
+from ghostos.abcd.utils import get_module_magic_matrix_providers
 from ghostos.scripts.cli.utils import (
     find_ghost_by_file_or_module,
 )
@@ -10,7 +10,7 @@ from ghostos_common.entity import get_entity
 
 def run_console_app(file_or_module: str):
     ghost_info, module, filename, is_temp = find_ghost_by_file_or_module(file_or_module)
-    providers = get_module_magic_shell_providers(module)
+    providers = get_module_magic_matrix_providers(module)
     ghostos = get_ghostos()
     ghost = get_entity(ghost_info.ghost, Ghost)
     app = ConsoleApp(ghostos=ghostos, ghost=ghost, username="", providers=providers)