+++ libs/container/RELEASES.md
@@ -2,6 +2,14 @@
 
 prepare uv and monorepo. 
 
+## 0.2.6
+
+fix bootstrapper lifecycle wrong.
+
+## 0.2.5
+
+add method `set_parent` to container.
+
 ## 0.2.4
 
 fix `get_caller_info` backtrace more than stack